Happy Number
isHappyNumber() determines whether a given number is happy or not.
If the number is greater than 0, calculate remainder rem by dividing the number with 10.
Calculate square of rem and add it to a variable sum.
Divide number by 10.
Repeat the steps from a to c till the sum of the square of all digits present in number has been calculated.
Finally, return the sum.