Disarium Number
ALGORITHM
STEP 1: calculateLength() counts the digits present in a number.
Use a while loop to check whether the number is not equal to 0.
Divide the number by 10 and increment the variable length by 1.
Return length.
STEP 2: Define and initialize variable num.
STEP 3: Make a copy of num by storing the value of num in n.
STEP 4: Using while loop calculates remainder rem repeatedly by dividing num with 10.
STEP 5: Calculate the value of rem raised to power its position, i.e., remlen and store the computed value in a variable sum.
STEP 6: Check whether the sum is equal to the number. If yes, then given number is Disarium number. Else, it is not a Disarium number.